# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText OpenAPI
  description: API for WASAText
  version: 1.0.0
security:
  - BearerAuth: [] 
...
paths:
...
  /conversations:
    get:
      tags:
        - Conversation
      summary: Retrieve user's conversations
      description: Gets a list of conversations the user is part of
      operationId: getMyConversations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user's conversations
          content:
            application/json:
              schema:
                description: Array of conversation objects representing the user's active and archived conversations.
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
                minItems: 0
                maxItems: 100
  /conversations/{conversationId}:
    get:
      tags:
        - Conversation
      summary: Retrieve a specific conversation
      description: Gets a list of all messages within a conversation, including text, photos, timestamps, and sender details.
      operationId: getConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            pattern: ^.*?$
            minLength: 1
            maxLength: 36
          description: Unique identifier for the conversation.
      responses:
        '200':
          description: Conversation details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: Invalid conversation ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages:             
    post:
      tags:
        - Message
      summary: Sending a message
      description: Sending a message in the specified chat.
      operationId: sendMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            description: Unique identifier for the conversation.
            pattern: ^.*?$
            minLength: 1
            maxLength: 36
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '201':
          description: Message successfully sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Message failed to send due to invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages/{messageId}/forward:
    post:
      tags:
        - Message
      summary: Forward a message
      description: Forwards an existing message to a specified conversation.
      operationId: forwardMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            description: Unique identifier for the conversation.
            pattern: ^.*?$
            minLength: 1
            maxLength: 36
        - name: messageId
          in: path
          required: true
          description: Unique identifier for the message.
          schema:
            type: string
            pattern: ^.*?$
            minLength: 1
            maxLength: 36
      requestBody:
        required: true
        description: Object containing the ID of the target conversation where the message will be forwarded.
        content:
          application/json:
            schema:
              type: object
              required:
                - targetConversationId
              properties:
                targetConversationId:
                  type: string
                  description: Unique identifier of the conversation to forward the message to.
                  pattern: ^.*?$
                  minLength: 1
                  maxLength: 36
      responses:
        '201':
          description: Message successfully forwarded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Failed to forward the message due to invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages/{messageId}:
    delete:
      tags:
        - Message
      summary: Delete a message
      description: Deletes a specific message within a conversation.
      operationId: deleteMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            description: Unique identifier for the conversation.
            pattern: ^.*?$
            minLength: 1
            maxLength: 36
        - name: messageId
          in: path
          required: true
          description: Unique identifier for the message.
          schema:
            type: string
            pattern: ^.*?$
            minLength: 1
            maxLength: 36
      responses:
        '204':
          description: Message deleted successfully.           
        '400':
          description: Failed to delete message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
      BearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT
        description: |
          Use a valid JWT token in the Authorization header to access protected endpoints.
  schemas:
...
    Username:
      type: string
      pattern: ^[a-zA-Z0-9_]+$
      minLength: 3
      maxLength: 16
      description: A username containing only letters, numbers, or underscores, with a maximum length of 16 characters
...
    Conversation:
      type: object
      description: Represents a user's conversation.
      properties:
        conversationId:
          type: string
          description: Unique identifier for the conversation.
          pattern: ^.*?$
          minLength: 1
          maxLength: 36
        displayName:
          type: string
          description: Display name of the conversation (e.g., group name or other user's name).
          minLength: 1
          maxLength: 100
        type:
          type: string
          enum: [direct, group]
          description: Indicates whether the conversation is a private chat or a group conversation.
        createdAt:
          type: string
          format: date-time
          description: Timestamp of when the conversation was created.
        members:
          type: array
          description: List of usernames in the conversation.
          items:
            $ref: '#/components/schemas/Username'
          minItems: 2
          maxItems: 100
        groupPhoto:
          type: string
          format: byte
          description: Base64-encoded image data
        lastMessage:
          $ref: '#/components/schemas/Message'
        messages:
          type: array
          description: List of messages in the conversation.
          items:
            $ref: '#/components/schemas/Message'
          minItems: 0
          maxItems: 100
    Message:
      type: object
      description: A message sent in a conversation.
      required:
        - id
        - sender
        - content
        - timestamp
        - message_status
        - reaction
        - attachment
      properties:
        id:
          type: string
          description: Unique identifier for the message.
        sender:
          type: object
          description: Information about the message sender.
          required:
            - id
            - username
          properties:
            id:
              type: string
              description: Unique identifier of the sender.
            username:
              $ref: '#/components/schemas/Username'
        content:
          type: object
          description: Content of the message.
          required:
            - type
            - value
          properties:
            type: 
              type: string
              enum: ['text', 'photo']
              description: Type of message (text or photo).
            value:
              type: string
              description: Message text or photo URL.
              pattern: ^.*?$
              maxLength: 500
              minLength: 1
        timestamp:
          type: string
          format: date-time
          description: Timestamp of when the message was sent.
        message_status:
          type: string
          enum: ['sent', 'delivered', 'read']
          description: Status of the message.
        reaction:
          type: array
          description: Reactions to this message.
          items:
            $ref: '#/components/schemas/Reaction'
          minItems: 0
          maxItems: 100
        attachments:
          type: array
          description: List of attachments.
          items:
            type: string
            format: byte
        forwarded_from:
          type: string
          description: ID of the user this message was forwarded from.
          pattern: ^.*?$
          minLength: 1
          maxLength: 36
    Reaction:
      type: object
      description: A reaction to a message.
      properties:
        emoji:
          type: string
          description: Emoji used for the reaction.
          pattern: ^.*?$
          minLength: 1
          maxLength: 10
        username:
          $ref: '#/components/schemas/Username'
...
    Error:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error details about why the request failed
...